requirements.txt
numpy
torch
torchvision
opencv-python
colorlog
matplotlib
Pillow
imageio
skimage
tqdm